requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.22
scikit-learn>=1.1
PyYAML>=6.0
matplotlib>=3.7
Pillow>=9.5
tqdm>=4.65
